Move get_features_targets to ml module
