+++ docs/reference/ml.rst
@@ -1,0 +1,5 @@
+Machine Learning
+=========
+
+.. automodule:: janitor.ml
+   :members:
