+++ pytest.ini
@@ -8,3 +8,4 @@ markers =
     finance: tests for finance
     utils: utility tests
     engineering: tests for engineering
+    ml: tests for machine learning
